requirements.txt
aiogram==2.23.1
fastapi==0.87.0
python-dotenv==0.21.0
uvicorn==0.19.0
telethon
openpyxl
psycopg2